Move ensure_filelike to utils
